test(broker): clean up tests
